+++ components/AddEvent.js
@@ -1,10 +1,12 @@
-import { StyleSheet, Text, TextInput, Modal, TouchableOpacity, View } from 'react-native';
+import { StyleSheet, TextInput, View, Alert } from 'react-native';
 import DateTimePicker from '@react-native-community/datetimepicker';
-import DatePicker from 'react-native-modern-datepicker';
 import React, { useState } from 'react';
 import addItem from './dataHandler';
 import RichTextEditor from './RichTextEditor';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
+import { get12HourTime } from './timeConvert';
+import OpacityButton from './OpacityButton';
+import { theme } from './Styles';
 
 /**
  * This view is used to add an event to the database
@@ -18,65 +20,26 @@ export default function AddEvent({ navigation, route }) {
   const [startTime, setStartTime] = useState('');
   const [title, setTitle] = useState('');
 
-  /**
-   * When the user selects a departure, hide the departure picker and set the departure to the selected
-   * time.
-   * @param selectedTime - The time that was selected by the user.
-   */
+  const departureButtonText = departure != '' ? get12HourTime(departure) : 'Departure Time';
+  const startTimeButtonText = startTime != '' ? get12HourTime(startTime) : 'Start Time';
+  const emptyFields = title == '' || description == '' || startTime == '';
+
   const onDepartureTimeChange = (event, selectedDate) => {
     setShowDeparturePicker(false);
     const date = new Date(selectedDate);
-    console.log(selectedDate);
     setDeparture(`${date.getHours()}:${date.getMinutes()}`);
   };
 
-  /**
-   * Takes a selectedDate as an argument; sets the
-   * startTime state to the selectedDate's hours and minutes, and then sets the showStartTimePicker state
-   * to false.
-   * @param selectedDate - The date that was selected by the user.
-   */
   const onStartTimeChange = (event, selectedDate) => {
     setShowStartTimePicker(false);
     const date = new Date(selectedDate);
-    console.log(selectedDate);
     setStartTime(`${date.getHours()}:${date.getMinutes().toString().padEnd(2, '0')}`);
   };
 
-  /**
-   * If the hour is less than 12 and greater than 0, it's AM. If the hour is 0, it's 12 AM. If the hour
-   * is greater than 12, it's PM. If the hour is 12, it's 12 PM.
-   * @param time - The time in 24 hour format.
-   * @returns A string with the hour, minutes, and AM/PM.
-   */
-  const get12HourTime = (time) => {
-    let hour = time.split(':')[0];
-    let am = 'AM';
-    if (hour < 12 && hour > 0) {
-      am = 'AM';
-    } else {
-      if (hour == 0) {
-        hour = 12;
-      } else {
-        hour = hour - 12;
-        if (hour == 0) {
-          hour = 12;
-        }
-        am = 'PM';
-      }
-    }
-    /* Returning a string with the hour, minutes, and AM/PM. */
-    return `${hour}:${time.split(':')[1]} ${am}`;
-  };
-
-  /**
-   * If the title, description, and departure are not empty, then add the item to the database and go back
-   * to the previous screen.
-   *
-   * If any of the fields are empty, then alert the user to fill out all fields.
-   */
   const saveAndExit = () => {
-    if (title !== '' && description !== '') {
+    if (emptyFields) {
+      Alert.alert('Oops', '\nPlease fill out all fields');
+    } else {
       addItem(
         'event',
         {
@@ -89,73 +52,79 @@ export default function AddEvent({ navigation, route }) {
         route.params.trace,
         () => navigation.goBack(null)
       );
-    } else {
-      alert('Please fill out all fields');
     }
   };
 
   return (
     <View style={styles.container}>
-      {/* This is a TextInput component that is used to get the title of the event. */}
       <TextInput
-        style={styles.textInput}
         placeholder={'Title'}
+        style={styles.textInput}
         onChangeText={(input) => setTitle(input)}
       />
 
-      {/* A button that sets the showStartTimePicker state to true when pressed. */}
-      <TouchableOpacity
-        style={styles.dateInput}
+      <OpacityButton
+        textStyle={theme.style.leftButton}
+        text={startTimeButtonText}
         onPress={() => {
-          if (showStartTimePicker == false) {
-            setShowStartTimePicker(true);
-          }
+          setShowStartTimePicker(true);
         }}
       >
-        <Text style={styles.dateInputText}>
-          <MaterialCommunityIcons name="clock" size={16} color="#5c5c5c" />
-          {startTime != '' ? get12HourTime(startTime) : ' Start Time'}
-        </Text>
-      </TouchableOpacity>
-      {/* A conditional statement that renders the DateTimePicker component if the
-      showStartTimePicker state is true. */}
+        <MaterialCommunityIcons
+          style={{ marginRight: 5 }}
+          color={theme.colors.text}
+          name="clock"
+          size={16}
+        />
+      </OpacityButton>
+
       {showStartTimePicker && (
         <DateTimePicker
-          testID="dateTimePicker"
           value={new Date('2000', '0', '01', '12', '00', '00', '00')}
-          mode={'time'}
-          is24Hour={false}
           onChange={onStartTimeChange}
+          testID="dateTimePicker"
+          is24Hour={false}
+          mode={'time'}
         />
       )}
-      {/* This is a button that sets the showPicker state to true when pressed. */}
-      <TouchableOpacity style={styles.dateInput} onPress={() => setShowDeparturePicker(true)}>
-        <Text style={styles.dateInputText}>
-          <MaterialCommunityIcons name="clock" size={16} color="#5c5c5c" />
-          {departure != '' ? get12HourTime(departure) : ' Departure Time'}
-        </Text>
-      </TouchableOpacity>
-      {/* A modal that is used to display the time picker for the event departure. */}
+
+      <OpacityButton
+        textStyle={theme.style.leftButton}
+        text={departureButtonText}
+        onPress={() => {
+          setShowDeparturePicker(true);
+        }}
+      >
+        <MaterialCommunityIcons
+          style={{ marginRight: 5 }}
+          color={theme.colors.text}
+          name="clock"
+          size={16}
+        />
+      </OpacityButton>
+
       {showDeparturePicker && (
         <DateTimePicker
-          testID="dateTimePicker"
           value={new Date('2000', '0', '01', '12', '00', '00', '00')}
-          mode={'time'}
-          is24Hour={false}
           onChange={onDepartureTimeChange}
+          testID="dateTimePicker"
+          is24Hour={false}
+          mode={'time'}
         />
       )}
-      {/* This is a TextInput component that is used to get the description of the event. */}
+
       <RichTextEditor
         placeholder={'Description'}
         callback={(response) => {
           setDescription(response);
         }}
       />
-      {/* A button that calls the done function when pressed. */}
-      <TouchableOpacity style={styles.done} onPress={saveAndExit}>
-        <Text style={styles.doneText}>Save Event</Text>
-      </TouchableOpacity>
+
+      <OpacityButton
+        buttonStyle={theme.style.bottomButton}
+        onPress={saveAndExit}
+        text={'Save Event'}
+      />
     </View>
   );
 }
@@ -163,8 +132,8 @@ export default function AddEvent({ navigation, route }) {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: 'white',
-    borderColor: '#67dfe8',
+    backgroundColor: theme.colors.white,
+    borderColor: theme.colors.accent,
     borderWidth: 5,
     padding: 15,
   },
@@ -186,8 +155,8 @@ const styles = StyleSheet.create({
 
   textInput: {
     fontSize: 16,
-    color: '#5c5c5c',
-    backgroundColor: '#f5f5f5ff',
+    color: theme.colors.text,
+    backgroundColor: theme.colors.itemColor,
     display: 'flex',
     flexDirection: 'row',
     justifyContent: 'space-between',
@@ -199,7 +168,7 @@ const styles = StyleSheet.create({
   },
 
   dateInput: {
-    backgroundColor: '#f5f5f5ff',
+    backgroundColor: theme.colors.itemColor,
     display: 'flex',
     flexDirection: 'row',
     justifyContent: 'space-between',
@@ -209,28 +178,9 @@ const styles = StyleSheet.create({
     marginVertical: 5,
   },
 
-  done: {
-    position: 'absolute',
-    bottom: 30,
-    width: '100%',
-    marginHorizontal: 20,
-    backgroundColor: '#f5f5f5ff',
-    paddingVertical: 15,
-    paddingHorizontal: 10,
-    borderRadius: 5,
-    marginVertical: 5,
-  },
-
   dateInputText: {
     fontSize: 16,
-    color: '#5c5c5c',
+    color: theme.colors.text,
     marginVertical: 5,
-  },
-
-  doneText: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    alignSelf: 'center',
-    color: '#5c5c5c',
   },
 });