+++ ChangeLog.md
@@ -1,5 +1,9 @@
 # ChangeLog
 
-- Added custom link previews via custom renderer in HTML renderer.
-- Fixed issues from Font-Scaling in timeline vertical layout
-- Fixed time selector modal reopening after selecting time.
+- made [custom button component](./components/OpacityButton.js)
+- added [theme file](./components/Styles.js)
+- cleaned up [AddEvent](./components/AddEvent.js)
+- cleaned up [AddTravel](./components/AddTravel.js)
+- cleaned up [AddTrip](./components/AddTrip.js)
+- cleaned up [RichTextEditor](./components/RichTextEditor.js)
+- extracted function get12HourTime(time) to [timeConvert](./components/timeConvert.js)